requirements.txt
numpy
scipy
librosa
torch
transformers
whisper
tensorflow
pandas
soundfile
pydub
scikit-learn
pyyaml